requirements.txt
environ-config
opencv-python-headless
uvicorn
fastapi
